Add experimental layout for Akkadian transliteration.
This has some similar characters, but is used for Babylonian
and Assyrian. Capital letters are necessary to represent
sign transliteration as distinct from normalized latin
representation of the language, so we need to do more
replacement of letters and dead keys for diacritics.

Need grave and accute accents too, but don't want to
clobber the brackets which are used to mark determinatives
and restorations.