Remove memrise font add-on.
The keyboard mappings were sharing a repository with a Firefox
add-on to tweak the fonts on the memrise.com flashcard website.
That might still be worth doing for these languages, since the
default fonts don't cover all the transliteration characters,
but this version was focussed on polytonic Greek, which has
developed further in another repository.

Remove the add-on files for now to make the repository more
clear.